+++ CHANGELOG.md
@@ -15,6 +15,7 @@ string beginning with the name of your module, anything you like after the first
 * File search refactored to only search for running modules
     * Makes search much faster when running with lots of files and limited modules
     * For example, if using `-m star` to only use the STAR module, all other file searches now skipped
+* File search now warns if an unrecognised search type is given
 
 ## [MultiQC v1.2](https://github.com/ewels/MultiQC/releases/tag/v1.2) - 2017-08-16